Update _index.md
We only allow MEMORY USAGE <key> command, and the rest of the flavors of MEMORY command are not allowed, like MEMORY STATS.
I believe this should be added to the list under FAQ "Are you fully compatible with open source Redis?" 

I have tried the commands in both Cloud and RS, and it is not allowed in both. Please check with PM if necessary.